+++ src/main/java/itstime/reflog/todolist/service/TodolistService.java
@@ -71,4 +71,6 @@ public class TodolistService {
 
 		todolistRepository.delete(todolist);
 	}
+
+
 }
